feat: Try using ServiceManager in two threads
Worked pretty well, but it can "spin" if there is no work to do, definitely
need to try select in crossbeam.

Make changes to work with the modified server_manager. These changes
were necessary as the the client and server don't automatically stop.
This also necessitated the changes to server_manager_1000 and I deleted
server_manager_multipele as I wasn't using it and have settled on
1000 loops as good enough.